Separate final fixes and cleanup for mainnet proposal branch
Rebase/cleanup mainnet proposal and get ready to deploy

Remove unnecessary testnet migrations

We had migrations to seed COMP on testnet (where it exists), but we now do it as part of deploy.
In general the correct principal for what should be a migration seems to be, only things that *must* be done as proposals.
In particular, changes made after a live deployment is handed over to governance.
Migrations should be deleted after the proposal passes and the PR with tests is merged.
This seems like the best policy, because migrations and proposals are expensive, even with potential improvements they combinatorially  multiply the number of scenarios we need to run.
So since we can seed rewards initially, we do.

A little more cleanup of utils