Mainnet deploy script and scenarios (#501)
This PR introduces mainnet deploy scripts and allows us to fully test hypothetical deployments of Comet using scenarios without having to first deploy the contracts on-chain. This is useful for our mainnet launch, where we want to test a hypothetical mainnet deployment before actually deploying. 

A few big changes in this PR:
- Adding a new deployment script and `configuration.json` for `mainnet` that allows us to test it using scenarios
- Supporting mainnet governance scenarios by having `fastExecuteGovernance` work for deployments with a real `Timelock` that require votes to pass proposals
- Allowing the `configuration.json` file to specify an optional `address` field for the base asset and each collateral asset. This is useful for chains that already have pre-existing tokens we want to support (unlike testnets, where we deploy our own set of tokens)
- Fixing scenarios that were failing for mainnet