Set RewardConfig during initial deployment (#512)
This PR updates the deploy function to set the `RewardConfig` in the `CometRewards` contract as part of the deployment process. 

Two optional fields have been added to `configuration.json` to specify the reward token for a certain deployment: `rewardToken` (token symbol) and `rewardTokenAddress`. If the rewards contract is to be deployed, then one of these two fields must be specified in the deployment's `configuration.json`.