Fix deployment manager contract cache and spidering issues
Adds interfaces for contracts we use in scens and deployments.
These artifacts are now used by spider, since we've had some trouble verifying the cloned contracts, and are afraid we may lose the ability to verify very old contracts, if we haven't already.
This also doesn't cache the artifacts spider reads locally in the contract cache, since that creates an extra step of having to purge the cache after changes are made to the artifacts, as well as creates a broken condition where if we first read the contract as an artifact, we won't be able to deploy it.
In all, this means we are guaranteed that everything in the contract cache is pulled from etherscan and deployable.

Finally, this gets rid of the invalidate/rebuild the contract cache for the deployment manager, since we can just directly use the results of spidering.
This is significantly faster, and avoids the problem that we would have rebuilding the contracts from disk without spidering now that the artifacts aren't kept in the cache.