Make getAssetIndex public, fix bumpSupplyCaps
This makes a contract change to configurator, to make `getAssetIndex` public.
It was useful to be able to call this directly when trying to debug these  issues and I think it ought to be public.

Somehow the upgrade path in bumpSupplyCaps was completely broken and had never been used (at least since adding the per-market configuration).
This fixes the call to updateSupplyCap and bypasses governance for this constraint helper to speed it up.

Also remove some whales, since we are voting with all of them and its unnecessary.