Rescue deployed contract roots
Fortunately the contracts deployed successfully through the workflow.
Unfortunately, they failed before verification (I believe an issue with seacrest during spidering).
This is now a bit of a problem because the contracts aren't verified and spider breaks trying to import the existing roots.
I constructed the roots.json manually from the logs and we can deal with the verification problem manually as well.