Fix proposal constraint and fetchQuery
Scen plugin should not depend on scens.
Errors should be thrown in fetchQuery to avoid misusage.
All open proposals should be executed, not just pending/active.

This increases the need for 'static constraints' as proposal constraint is very expensive.
This also hacks in a constant for timelock buf which should really be computed.