Redeploy comet with different USDC
Keep the rewards contract since it has COMP already, though need to pass a proposal to set rewards config later.

Also, the USDC base is not one we deployed and we cannot mint, so we should be ensure to grant the fauceteer.

Re-deploy Goerli with other USDC

Fix USDC address, which cannot be minted