Separate whales for testnet
For prior testnets, we completely controlled the COMP allocation, so we were able to delegate to the top mainnet whale when cloning gov.
For new goerli, we are leveraging an existing gov installation, so we need to know about some external COMP whales.
The current mechanism is sort of a compromise to keep complexity down and prevent scens from being even slower.
We shouldn't necessarily need any sets of whales beyond 'mainnet' and 'testnet'.