Automatically include excess balance in collateral reserves
Also adds a reserves fn which it checks when collateral is bought.

Note: we could add a fn to convert any (prior) collateral balance of the contract into reserves by clearing it out and removing from total.
However, collateral reserves are currently 0, and we expect them to remain 0 if liquidation is working properly.
If somehow collateral is acquired and held by the protocol (i.e. from an absorb) before this patch gets applied, the protocol will simply hold that balance as a supplier, until governance decides to move it, for which a fn can always be added (at that time) to assist.