Patches for hardhat tests
This gets rid of the dependency on `ethereum-waffle` and uses a patched version of `hardhat-chai-matchers`, built from the same branch as our (updated) patched `hardhat`.
It attempts to preserve the interface we were using to match errors as much as possible, but there were a few tests that needed to be updated.
Notably, the matcher we are currently using doesn't have the ability to match custom error arguments.
We might want to switch to matching with `revertedWithCustomError` if there are improvements and/or this stabilizes in the future, but for now attempted to keep minimal changes to tests with the matcher changes.

We also move the Comet errors into the interface, in order to make things simpler/more consistent for the future, even though its not totally necessary right now.