Price feeds for WETH deployment + Bulker changes for OZ audit (#625)
This PR implements and modifies price feeds to support the upcoming WETH deployment. The favored plan so far is to use ETH-denominated price feeds as opposed to USD price feeds, but stick with using 8 decimals for prices to avoid having to change the `Comet` and `Configurator` implementations. 

This would require: 
- A new wrapper price feed (`ScalingPriceFeed.sol`) that scales prices up or down to 8 decimals
- A new `ConstantPriceFeed` that always returns 1e8 for the `WETH` base asset, since should always hold a 1:1 value with ETH
- Modifications to the `WstETHPriceFeed` to return prices in terms of ETH instead of USD

This is an alternative approach to #626, which is a more complex change but could be a better long-term solution.

*Note: This PR also now contains the changes from #634 and #635, which address some suggestions made by OZ for their audit of `WstETHPriceFeed` and `Bulker`.*