Support multiple connections to Seacrest in Enact (#692)
Currently, running `Enact` in CI using Seacrest will fail when multiple networks are used (e.g. for cross-chain governance proposals that read from Polygon and write to mainnet). This is because Seacrest's connection with WalletConnect is for a specific chain, so trying to access a different chain will result in a `could not detect network` error.

This change supports making multiple connections to Seacrest/WalletConnect, all from a single phone. We introduce optional `GOV_NETWORK` and `GOV_NETWORK_PROVIDER` env variables to override the provider for a specific chain when creating the HREs. The `Enact` CI job determines the `GOV_NETWORK` based on what the original selected network is.