Diff checker for state changes after running a migration (#729)
This adds a diff checker to check for state changes after a migration is run. This allows us to more comprehensively check for state changes because, right now, our migration verification logic will not catch any unintentional state changes. For example, in a proposal that only touches collateral factors, we'll check that `collateralFactor == some new amount`, but not that `liquidationFactor == old amount`. We could also potentially generate a report of all the state changes for easier viewing.

In order to isolate the state changes of each migration, we get rid of the `VerifyMigrationConstraint` and move the verify checks directly into the `ProposalConstraint` so that verification is called right after a proposal is executed. This way, we can verify the state changes of a specific migration without having the diffs be affected by other migrations or pending proposals.