QoL improvement for verifying contracts when deploying (#741)
This PR improves the deploy step by allowing us to recover the `VerifyArgs` if a deploy job fails mid-deploy. This allows us to easily resume verification of contracts that were deployed but not yet verified. We also upgrade `hardhat-etherscan` to support verifying contracts on L2s.

The changes in this PR are extracted from #728, which currently lives on the Optimism branch. I'm moving them to main since these are generic quality of life improvements that should be used on other branches as well, such as the Arbitrum one.