Skip utilization scenarios for Goerli USDC (#750)
Someone supplied 100bn USDC to the Goerli USDC market, which caused 2 scenarios using the utilization constraint to fail because it became impossible to achieve a target utilization with so much supply in the market. This PR skips those 2 scenarios.

Another solution is to enhance the utilization constraint such that it can also withdraw base supplied by whales to achieve the target utilization.