Use Sleuth to Bulk Liquidity Checks
This patch uses Sleuth to bulk the liquidity checks on a given network. This should drop us from an O(n) query to an O(1) query. There are a number of other improvements we can make to drop this further, e.g. by using block_number as returned from Sleuth instead of constantly querying the block number; plus we should convert to using StaticJsonRpcProvider as another way to drop our query count. But that all aside, this is the first and biggest improvement we could make.

Note: I'm a bit weary that we might ask for so many addresses' liquidities in one query that it runs out of gas (e.g. max block gas) and then doesn't work. I would propose that we change this to using a partial scan each time (e.g. 100-200 addresses at random) and we would pretty quickly converge on all addresses (that is, allowing us to make multiple smaller queries per block, but allowing this to be somewhat random). If we sliced the pool for each query and reshuffled on each new block, we would be close to guaranteeing that we will check all addresses within one or two blocks.

That is, you basically run:

```
given allAddresses;

let blockNumber = currentBlockNumber;

while(true) {
  let random = allAddresses.shuffle();
  while (random.length > 0) {
    let [blockNumber, checks] = check(random.pop(0, 100)); // Pop 100 addresses and check them
    // do something with checks
    if (blockNumber > currentBlockNumber) {
      currentBlockNumber = blockNumber;
      break; // We're on a new block, reshuffle and start over
    }
    sleep(5000); // Wait a few moments
    // ... and then check more addresses from this shuffle
  }

  // Reshuffle and start over, given we're on a new block
}
```

That's a little different, but basically it should converge on checking all addresses with high probability.