Base Goerli cWETHv3 deploy (#751)
This PR introduces the cWETHv3 deployment on Base Goerli as well as the proposal to initialize the market there.

When initializing the cWETHv3 market, we bridge over some WETH reserves. However, Optimism/Base does not support bridging WETH directly, so we bridge over ETH instead and wrap it before sending it to cWETHv3. I had to add some logic to our mock relayer code to support bridging of ETH.

* Deploy script for cWETHv3 on Base Goerli

* Create proposal to initialize cWETHv3 and bridge ETH as reserves

* Run base-goerli-weth in scenarios CI

* Fix pricefeeds path after rebase

* Address PR comments

* Modified deployment roots from GitHub Actions

* Update proposal

* Modified migration from GitHub Actions

---------

Co-authored-by: GitHub Actions Bot <>