Make updateBasePrincipal virtual so that it can be overridden (#749)
My team received a grant from CGP 2.0 to add Flexible Voting support to Compound V3.

In order to implement our extension we need to be able to determine, for any given block:

how much principal a given account has supplied up to that time, and
how much total principal has been supplied to Comet up to that time
Our belief is that the easiest way to do this involves overriding the Comet.updateBasePrincipal function to store a couple checkpoints. You can see the PR for this work in ScopeLift/flexible-voting#46, and the specific overrides we need to make here. Since updateBasePrincipal is the single way that principal balances are updated by the system, being able to add checkpointing to it greatly simplifies our task.

Thanks in advance!